+++ include/scheduler/services/ipc_service.h
@@ -20,9 +20,9 @@ typedef enum {
 
 /** @brief Minimal ISR-captured IPC metadata for deferred handling. */
 typedef struct {
-    uint8_t endpoint;
-    uint8_t slot;
-    uint8_t irq_tag;
+    uint16_t endpoint;
+    uint16_t slot;
+    uint16_t irq_tag;
     uint16_t detail_flags;
 } sch_ipc_irq_record_t;
 
@@ -35,7 +35,7 @@ typedef enum {
 
 /** @brief HAL extension points for platform-specific IPC interactions. */
 typedef struct {
-    void (*ack_irq)(void *hal_ctx, sch_ipc_irq_tag_t tag, uint8_t endpoint);
+    void (*ack_irq)(void *hal_ctx, sch_ipc_irq_tag_t tag, uint16_t endpoint);
     bool (*service_notify)(void *hal_ctx, const sch_ipc_irq_record_t *record);
     bool (*service_ack)(void *hal_ctx, const sch_ipc_irq_record_t *record);
     bool (*service_fault)(void *hal_ctx, const sch_ipc_irq_record_t *record);
@@ -59,12 +59,12 @@ bool sch_ipc_service_init(
     sch_overflow_mode_t overflow_mode);
 
 void sch_ipc_isr_notify(
-    sch_ipc_service_t *service, uint8_t endpoint, uint8_t slot, uint16_t payload_flags);
+    sch_ipc_service_t *service, uint16_t endpoint, uint16_t slot, uint16_t payload_flags);
 
-void sch_ipc_isr_ack(sch_ipc_service_t *service, uint8_t endpoint, uint8_t slot, uint16_t ack_flags);
+void sch_ipc_isr_ack(sch_ipc_service_t *service, uint16_t endpoint, uint16_t slot, uint16_t ack_flags);
 
 void sch_ipc_isr_fault(
-    sch_ipc_service_t *service, uint8_t endpoint, uint16_t fault_flags, uint8_t source_slot);
+    sch_ipc_service_t *service, uint16_t endpoint, uint16_t fault_flags, uint16_t source_slot);
 
 void sch_ipc_service_run(void *ctx);
